requirements.txt
Django==3.1.2
django-ckeditor==6.0.0
django-model-utils==4.1.1